+++ cx-portal/src/components/pages/ServiceReleaseProcess/index.tsx
@@ -23,8 +23,8 @@ import { useNavigate } from 'react-router-dom'
 import { PAGES } from 'types/Constants'
 import { setCurrentActiveStep } from 'features/appManagement/slice'
 import { useDispatch } from 'react-redux'
-import { setAppId, setAppStatus } from 'features/appManagement/actions'
-import { initialState } from 'features/appManagement/types'
+import { setServiceId, setServiceStatus } from 'features/appManagement/actions'
+import { initialServiceState } from 'features/appManagement/types'
 import { ReleaseProcess } from 'components/shared/basic/ReleaseProcess'
 
 export default function ServiceReleaseProcess() {
@@ -62,8 +62,8 @@ export default function ServiceReleaseProcess() {
   const onStartServiceClick = () => {
     navigate(`/${PAGES.SERVICERELEASEPROCESS}/form`)
     dispatch(setCurrentActiveStep())
-    dispatch(setAppId(''))
-    dispatch(setAppStatus(initialState.appStatusData))
+    dispatch(setServiceId(''))
+    dispatch(setServiceStatus(initialServiceState.serviceStatusData))
   }
 
   const onOverviewServiceButton = () => {}