+++ cx-portal/src/components/shared/basic/ReleaseProcess/TechnicalIntegration/index.tsx
@@ -41,7 +41,7 @@ import {
   useUpdateRoleDataMutation,
 } from 'features/appManagement/apiSlice'
 import { setAppStatus } from 'features/appManagement/actions'
-import SnackbarNotificationWithButtons from '../SnackbarNotificationWithButtons'
+import SnackbarNotificationWithButtons from '../components/SnackbarNotificationWithButtons'
 
 export default function TechnicalIntegration() {
   const { t } = useTranslation()
@@ -95,7 +95,7 @@ export default function TechnicalIntegration() {
   })
 
   useEffect(() => {
-    dispatch(setAppStatus(fetchAppStatus))
+    if (fetchAppStatus) dispatch(setAppStatus(fetchAppStatus))
   }, [dispatch, fetchAppStatus])
 
   const onIntegrationSubmit = async (data: any, buttonLabel: string) => {