+++ cx-portal/src/features/appManagement/apiSlice.ts
@@ -20,6 +20,7 @@
 
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
 import { apiBaseQuery } from 'utils/rtkUtil'
+import { AppStatusDataState, ServiceStatusDataState } from './types'
 
 export type useCasesItem = {
   useCaseId: string
@@ -49,6 +50,20 @@ export type CreateAppStep1Item = {
   supportedLanguageCodes: string[]
   price: string
   privacyPolicies: string[]
+}
+
+export type CreateServiceStep1Item = {
+  title?: string
+  price?: string | null
+  leadPictureUri?: string
+  descriptions?: {
+    languageCode: string
+    longDescription: string
+    shortDescription: string
+  }[]
+  privacyPolicies?: string[]
+  salesManager?: string | null
+  serviceTypeIds?: string[]
 }
 
 export type ImageType = {
@@ -89,7 +104,7 @@ export type AgreementType = {
 
 export type AgreementStatusType = {
   agreementId: string
-  consentStatus: string | boolean
+  consentStatus: ConsentStatusEnum
 }
 
 export type ConsentType = {
@@ -104,6 +119,11 @@ export type UpdateAgreementConsentType = {
 export type saveAppType = {
   appId: string
   body: CreateAppStep1Item
+}
+
+export type createServiceType = {
+  id: string
+  body: CreateServiceStep1Item
 }
 
 export type salesManagerType = {
@@ -147,6 +167,31 @@ export type updateRolePayload = {
 export type DocumentRequestData = {
   appId: string
   documentId: string
+}
+
+export interface ServiceTypeIdsType {
+  serviceTypeId: number
+  name: string
+}
+
+export enum ConsentStatusEnum {
+  ACTIVE = 'ACTIVE',
+  INACTIVE = 'INACTIVE',
+  PENDING = 'PENDING',
+}
+
+export enum DocumentTypeId {
+  CX_FRAME_CONTRACT = 'CX_FRAME_CONTRACT',
+  COMMERCIAL_REGISTER_EXTRACT = 'COMMERCIAL_REGISTER_EXTRACT',
+  APP_CONTRACT = 'APP_CONTRACT',
+  CONFORMITY_APPROVAL_REGISTRATION = 'CONFORMITY_APPROVAL_REGISTRATION',
+  ADDITIONAL_DETAILS = 'ADDITIONAL_DETAILS',
+  APP_LEADIMAGE = 'APP_LEADIMAGE',
+  APP_IMAGE = 'APP_IMAGE',
+  SELF_DESCRIPTION = 'SELF_DESCRIPTION',
+  APP_TECHNICAL_INFORMATION = 'APP_TECHNICAL_INFORMATION',
+  CONFORMITY_APPROVAL_CONNECTOR = 'CONFORMITY_APPROVAL_CONNECTOR',
+  CONFORMITY_APPROVAL_BUSINESS_APPS = 'CONFORMITY_APPROVAL_BUSINESS_APPS',
 }
 
 export const apiSlice = createApi({
@@ -202,7 +247,7 @@ export const apiSlice = createApi({
           : { error: response.error }
       },
     }),
-    fetchAppStatus: builder.query<any, string>({
+    fetchAppStatus: builder.query<AppStatusDataState, string>({
       query: (appId) => `api/apps/appreleaseprocess/${appId}/appStatus`,
     }),
     fetchAgreementData: builder.query<AgreementType[], void>({
@@ -270,6 +315,66 @@ export const apiSlice = createApi({
         }),
       }),
     }),
+    createService: builder.mutation<void, createServiceType>({
+      query: (data) => ({
+        url: `/api/services/addService`,
+        method: 'POST',
+        body: data.body,
+      }),
+    }),
+    saveService: builder.mutation<void, createServiceType>({
+      query: (data) => ({
+        url: `/api/services/${data.id}`,
+        method: 'PUT',
+        body: data.body,
+      }),
+    }),
+    fetchServiceStatus: builder.query<ServiceStatusDataState, string>({
+      query: (id) => `api/services/${id}`,
+    }),
+    fetchServiceTypeIds: builder.query<ServiceTypeIdsType, void>({
+      query: () => `/api/services/servicerelease/serviceTypes`,
+    }),
+    updateServiceAgreementConsents: builder.mutation<
+      void,
+      UpdateAgreementConsentType
+    >({
+      query: (data: UpdateAgreementConsentType) => ({
+        url: `/api/services/${data.appId}/serviceAgreementConsent`,
+        method: 'POST',
+        body: data.body,
+      }),
+    }),
+    fetchServiceAgreementData: builder.query<AgreementType[], void>({
+      query: () => `api/services/servicerelease/agreementData`,
+    }),
+    fetchServiceConsentData: builder.query<ConsentType, string>({
+      query: (appId: string) => `/api/services/servicerelease/consent/${appId}`,
+    }),
+    updateServiceDocumentUpload: builder.mutation({
+      async queryFn(
+        data: {
+          appId: string
+          documentTypeId: DocumentTypeId
+          body: { file: File }
+        },
+        _queryApi,
+        _extraOptions,
+        fetchWithBaseQuery
+      ) {
+        const formData = new FormData()
+        formData.append('document', data.body.file)
+
+        const response = await fetchWithBaseQuery({
+          url: `/api/services/updateservicedoc/${data.appId}/documentType/${data.documentTypeId}/documents`,
+          method: 'PUT',
+          body: formData,
+        })
+        return response.data
+          ? { data: response.data }
+          : { error: response.error }
+      },
+    }),
   }),
 })
 
@@ -292,4 +397,12 @@ export const {
   useDeleteRolesMutation,
   useDeleteDocumentMutation,
   useFetchNewDocumentByIdMutation,
+  useSaveServiceMutation,
+  useCreateServiceMutation,
+  useFetchServiceStatusQuery,
+  useFetchServiceTypeIdsQuery,
+  useUpdateServiceAgreementConsentsMutation,
+  useFetchServiceAgreementDataQuery,
+  useFetchServiceConsentDataQuery,
+  useUpdateServiceDocumentUploadMutation,
 } = apiSlice